Modularize slicedimage/io
Move URL path methods to slicedimage/url/path.py
Move URL resolve methods to slicedimage/url/resolve.py
Updated tests to reflect new organization.
