+++ Makefile
@@ -10,7 +10,13 @@ test: lint
 $(test_srcs): %.py :
 	python -m unittest $(subst /,.,$*); \
 
-lint:
-	flake8 $(MODULES)
+lint:   lint-non-init lint-init
+
+# TODO: (ttung) remove W504
+lint-non-init:
+	flake8 --ignore 'E252, E301, E302, E305, E401, W503, W504, E731, F811' --exclude='*__init__.py' $(MODULES)
+
+lint-init:
+	flake8 --ignore 'E252, E301, E302, E305, E401, F401, W503, W504, E731, F811' --filename='*__init__.py' $(MODULES)
 
 .PHONY : $(test_srcs)
