+++ slicedimage/io/__init__.py
@@ -1,0 +1,3 @@
+from ._base import Reader, VERSIONS, Writer
+from ._v0_0_0 import v0_0_0
+from ._v0_1_0 import v0_1_0
