+++ tests/io_/v0_1_0/test_missing_shape.py
@@ -9,7 +9,7 @@ import numpy as np
 
 import slicedimage
 from slicedimage._dimensions import DimensionNames
-from slicedimage.io import TileKeys, TileSetKeys
+from slicedimage.io._keys import TileKeys, TileSetKeys
 
 baseurl = "file://{}".format(os.path.abspath(os.path.dirname(__file__)))
